backend/agents/tools.py
from crewai_tools import tool
import requests
import time

# Exemplo de ferramenta para simular busca de preço de mercado
@tool("Market Price Search Tool")
def market_price_search(product_name: str) -> str:
    """
    Search for market prices of a given product.
    Simulates fetching data from an external market API or database.
    """
    print(f"\n> Executing Market Price Search for: {product_name}")
    time.sleep(2) # Simula atraso da API
    if "computador" in product_name.lower():
        return "Preço de mercado atual para computador: R$ 3500.00 (baseado em média de 3 fornecedores)."
    return "Preço de mercado não encontrado para este produto."

# Exemplo de ferramenta para simular análise de especificações online
@tool("Specification Analysis Tool")
def analyze_specs_online(spec_details: str) -> str:
    """
    Analyzes technical specifications by comparing them with industry standards online.
    Simulates searching a database of standards or a public knowledge base.
    """
    print(f"\n> Executing Specification Analysis for: {spec_details}")
    time.sleep(3)
    if "RAM 16GB" in spec_details:
        return "Especificação de RAM 16GB está de acordo com os padrões atuais para uso corporativo médio. Verificar compatibilidade com outros componentes."
    return "Análise de especificações técnicas: Detalhes insuficientes ou fora do padrão."

# Você criaria mais ferramentas aqui, por exemplo:
# @tool("Database Lookup Tool")
# def database_lookup(query: str) -> str:
#    """Accesses the internal PostgreSQL database to retrieve specific data related to past bids."""
#    # Lógica para conectar e consultar o DB (usar SQLAlchemy aqui)
#    return "Dados do DB..."