backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary # Driver PostgreSQL
pydantic
crewai
crewai-tools
python-dotenv # Para variáveis de ambiente